Efficient concat method and ID to name conversion
Now when new data is downloaded it will be concatenated to the existing one unless the parameter 'force_update' is set to True. In that case the data will all be concatenated. This is usefull in cases some downloaded games had corrupt data or the names of the players were wrong, for example.

Also added the names of the runes, champions, summoner spells and items. I might forget other IDs but I cant recall right now.